Changed `send` and `seal` to be `IO` actions
The `STM` requirement for `send` actually makes most use cases more difficult.
For example, to convert an action of type `a -> IO Bool` to an `Output`, you
have to use `spawn'`, which is overkill.  There seems to be no benefit to
preserving `STM` and there are many benefits to simplifying it to `IO`.